Put y Delete modificados para cabio de nombre y resticcion de eliminar

+++ src/app/groups/groups.controller.ts
@@ -14,6 +14,7 @@ import {
   import { GroupsEntity } from './groups.entity';
   import { TeachersService } from '../teachers/teachers.service';
   import { CreateGroupDto } from './dto/create-group.dto';
+  import { UpdateGroupDto } from './dto/update-group-dto';
  
   
 @Controller('groups')
@@ -39,8 +40,8 @@ export class GroupsController {
     if (!teacher) {
       throw new NotFoundException('No existe maestro');
     }
-    const nombreGrupo = `${group.schedule} ${teacher.firstName}`;
-    const newGroup = new GroupsEntity(group.schedule, nombreGrupo, teacher);
+    const nombreGrupo = `${group.day} ${group.schedule} ${teacher.firstName}`;
+    const newGroup = new GroupsEntity( group.schedule, nombreGrupo, teacher, group.day);
     const confirmGroup = await this.groupsService.findOneBy({
       name: newGroup.name,
     });
@@ -50,14 +51,40 @@ export class GroupsController {
     return this.groupsService.create(newGroup);
   }
   @Put(':id')
-  update(
+  async update(
     @Param('id') id: number,
     @Body() group: GroupsEntity,
   ): Promise<GroupsEntity> {
-    return this.groupsService.update(id, group);
+    const existingGroup = await this.groupsService.findOne(id);
+
+    if (!existingGroup) {
+      throw new NotFoundException('Grupo no encontrado');
+    }
+
+    existingGroup.schedule = group.schedule;
+    existingGroup.day = group.day;
+
+  const teacher = await this.teacherService.findOneById(group.teacher);
+  if (!teacher) {
+    throw new NotFoundException('No existe maestro');
+  }
+  existingGroup.name = `${group.day} ${group.schedule} ${teacher.firstName}`;
+
+    return this.groupsService.update(id, existingGroup);
   }
   @Delete(':id')
-  remove(@Param('id') id: string): Promise<void> {
-    return this.groupsService.remove(id);
+  async remove(@Param('id') id: string): Promise<void> {
+    const group = await this.groupsService.findOneWithRelations(id);
+  
+    if (!group) {
+      throw new NotFoundException('Grupo no encontrado');
+    }
+  
+    if (group.students && group.students.length > 0) {
+      throw new ConflictException('No se puede eliminar un grupo con alumnos');
+    }
+  
+    // Si no hay alumnos asociados, se puede eliminar el grupo
+    await this.groupsService.remove(id);
   }
 }
